requirements.txt
discord.py
asyncio
aiohttp
os
random
time
socket
psutil
asyncpraw
python-dotenv
pyfiglet